fraud-detection-pipeline/requirements.txt
# Python dependencies for Fraud Detection Pipeline
# Lambda Architecture with Batch, Speed, and Serving Layers

# =============================================================================
# Core Dependencies
# =============================================================================

# Apache Spark
pyspark==3.5.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# MongoDB
pymongo==4.6.0

# =============================================================================
# Data Processing
# =============================================================================

pandas==2.1.0
numpy==1.26.0
pyarrow==14.0.0

# =============================================================================
# Dashboard - Dash (Plotly)
# =============================================================================

dash==2.14.0
dash-bootstrap-components==1.5.0
plotly==5.18.0

# =============================================================================
# Monitoring - Prometheus
# =============================================================================

prometheus-client==0.19.0

# =============================================================================
# Web Framework (Optional - for REST API)
# =============================================================================

flask>=1.0.4,<2.3.0
flask-cors>=3.0.0

# =============================================================================
# Data Generation
# =============================================================================

faker==21.0.0

# =============================================================================
# Testing
# =============================================================================

pytest==7.4.0
pytest-cov==4.1.0

# =============================================================================
# Utilities
# =============================================================================

python-dateutil==2.8.2
requests==2.31.0
pyyaml==6.0.1

# =============================================================================
# Optional: ML Enhancement
# =============================================================================

# scikit-learn==1.3.0  # For additional ML algorithms
# mlflow==2.9.0        # For ML experiment tracking

# =============================================================================
# Development
# =============================================================================

black==23.11.0
flake8==6.1.0
mypy==1.7.0
